// src/types/nutrition.ts
export type Nutrients = {
  kcal: number;
  proteinG: number;
  fatG: number;
  carbG: number;
  fiberG: number;
  calciumMg: number;
  phosphorusMg: number;
  sodiumMg: number;
  potassiumMg: number;
  zincMg: number;
  copperMg: number;
  iodineUg: number;
  vitaminDUg: number;
  vitaminEMg: number;
  epaMg: number;
  dhaMg: number;
};

export const emptyNutrients: Nutrients = {
  kcal: 0,
  proteinG: 0,
  fatG: 0,
  carbG: 0,
  fiberG: 0,
  calciumMg: 0,
  phosphorusMg: 0,
  sodiumMg: 0,
  potassiumMg: 0,
  zincMg: 0,
  copperMg: 0,
  iodineUg: 0,
  vitaminDUg: 0,
  vitaminEMg: 0,
  epaMg: 0,
  dhaMg: 0
};